Added plugin handler for GET,POST.
For GET it validates roles, config and responds with the ogr2ogr help page
(useless).
For POST it validates roles, config, content type and expects a JSON object
with the extension as key and base64-encoded content as value:
{
  "shx": "AAAnCgAAAAAAAAAAAAAAAAAAAAAAAA etc. etc. AAAAAAAAAAAAAAAAADIAAAAK",
  "shp": "AAAnCgAAAAAAAAAAAAAAAAA etc. etc. KUVpslRBUsnDnUN8VUE=",
  "qpj": "UFJPSkNTWyJHZXJtYW55X1 etc. etc. FPUklUWVsiRVBTRyIsIjMxNDk1Il1dCg==",
  "prj": "UFJPSkNTWyJHZXJtYW55X1 etc. etc. 1d",
  "dbf": "A18HGgEAAABhAKEAAAAAAAA etc. etc. etc. AgICAgICAgICAgICAgICAgICAg"
}
It then writes every file in a temp dir as "remove-me.<extension>", runs

$ ogr2ogr -f GeoJSON <tmp-dir>/remove-me.geojson <tmp-dir>/remove-me.shp

and returns the contents of <tmp-dir>/remove-me.geojson.