Implemented use of GDAL_DATA and fallback_crs.
In a first step, remove-me.geojson is written by ogr2ogr and the plugin checks
if it also contains a 'crs' property, in which case the GeoJSON can be returned
to the client.
If the property is missing, we provide GDAL_DATA in env and transform all
coordinates to the given fallback_crs, to ensure that we end up with the crs
property.